Referencing BlockchainProcessing Nuget CSV Writing Cleaned Up CSV Caching Cleaned Up

+++ Nethereum.BlockchainStore.AzureTables/Repositories/BlockRepository.cs
@@ -83,13 +83,13 @@ namespace Nethereum.BlockchainStore.AzureTables.Repositories
             return blockOutput;
         }
 
-        public async Task<long> GetMaxBlockNumberAsync()
+        public async Task<ulong> GetMaxBlockNumberAsync()
         {
             await _lock.WaitAsync();
             try
             {
                 await InitialiseMaxBlock();
-                return _maxBlockCounter.Value;
+                return (ulong)_maxBlockCounter.Value;
             }
             finally
             {
